DBB、DBC、DBD　テーブル最新化

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/JutakuKaishuRiyushoTesuryoShukei.java
@@ -251,7 +251,7 @@ public class JutakuKaishuRiyushoTesuryoShukei
      *
      * @return 口座ID
      */
-    public Decimal get口座ID() {
+    public long get口座ID() {
         return entity.getKozaID();
     }
